pyproject.toml
[project]
name = "taipy-tools"
version = "0.1.0"
description = "Taipy application with small tools such as a UUID generator, a QR Code generator and a video to GIF converter."
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "ffmpeg-python==0.2.0",
    "segno==1.6.6",
    "taipy==4.1.0",
    "uuid-utils==0.11.0",
]
